# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "KLSI 4.0 API",
    "description": "Assessment engine and analytics API for KLSI 4.0",
    "version": "4.0.0"
  },
  "paths": {
...
    "/users/me": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "Get Me",
        "operationId": "get_me_users_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserOut"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/users/me/achievements": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "List User Achievements",
        "operationId": "list_user_achievements_users_me_achievements_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/UserAchievementOut"
                  },
                  "type": "array",
                  "title": "Response List User Achievements Users Me Achievements Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
    "/telemetry/batch": {
      "post": {
        "tags": [
          "telemetry"
        ],
        "summary": "Batch Telemetry",
        "description": "Accept batched telemetry events for async processing.\n\n**Scalability Benefits:**\n- Reduces request volume by 90%+ through client-side batching\n- Uses Beacon API for fire-and-forget delivery\n- Processes events asynchronously (non-blocking)\n- Handles up to 1000 events per batch\n\n**Usage:**\n```javascript\nconst batch = {\n  session_id: \"abc123\",\n  events: [\n    {type: \"mouse\", timestamp_ms: 1234567890, payload: {x: 100, y: 200}},\n    {type: \"scroll\", timestamp_ms: 1234567891, payload: {scrollY: 500}}\n  ]\n};\n\n// Using Beacon API (recommended)\nnavigator.sendBeacon('/telemetry/batch', JSON.stringify(batch));\n\n// Or fetch with keepalive\nfetch('/telemetry/batch', {\n  method: 'POST',\n  headers: {'Content-Type': 'application/json'},\n  body: JSON.stringify(batch),\n  keepalive: true\n});\n```\n\n**Returns:** 202 Accepted (processing in background)",
        "operationId": "batch_telemetry_telemetry_batch_post",
        "parameters": [
          {
            "name": "Content-Length",
            "in": "header",
            "required": true,
            "schema": {
              "type": "integer",
              "exclusiveMaximum": 1000000,
              "title": "Content-Length"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TelemetryBatch"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/telemetry/replay-events": {
      "post": {
        "tags": [
          "telemetry"
        ],
        "summary": "Record Replay Events",
        "description": "Record session replay events for debugging.\n\nStores events in a JSONL file for later playback/analysis.",
        "operationId": "record_replay_events_telemetry_replay_events_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReplayEventBatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BadgeOut": {
        "properties": {
          "slug": {
            "type": "string",
            "title": "Slug"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "rarity": {
            "type": "string",
            "title": "Rarity"
          }
        },
        "type": "object",
        "required": [
          "slug",
          "name",
          "rarity"
        ],
        "title": "BadgeOut"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ReplayEvent": {
        "properties": {
          "type": {
            "type": "string",
            "title": "Type"
          },
          "payload": {
            "type": "object",
            "title": "Payload"
          },
          "timestampMs": {
            "type": "integer",
            "title": "Timestampms"
          }
        },
        "type": "object",
        "required": [
          "type",
          "payload",
          "timestampMs"
        ],
        "title": "ReplayEvent",
        "description": "Event captured for session replay."
      },
      "ReplayEventBatch": {
        "properties": {
          "sessionId": {
            "type": "integer",
            "title": "Sessionid"
          },
          "events": {
            "items": {
              "$ref": "#/components/schemas/ReplayEvent"
            },
            "type": "array",
            "title": "Events"
          }
        },
        "type": "object",
        "required": [
          "sessionId",
          "events"
        ],
        "title": "ReplayEventBatch",
        "description": "Batch of replay events."
      },
...
      "Role": {
        "type": "string",
        "enum": [
          "MAHASISWA",
          "MEDIATOR"
        ],
        "title": "Role"
      },
...
      "TelemetryBatch": {
        "properties": {
          "session_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Id",
            "description": "Optional session ID for correlation"
          },
          "events": {
            "items": {
              "$ref": "#/components/schemas/TelemetryEvent"
            },
            "type": "array",
            "maxItems": 1000,
            "title": "Events",
            "description": "Events to process"
          }
        },
        "type": "object",
        "required": [
          "events"
        ],
        "title": "TelemetryBatch",
        "description": "Batch of telemetry events for efficient processing."
      },
      "TelemetryEvent": {
        "properties": {
          "type": {
            "type": "string",
            "title": "Type",
            "description": "Event type: mouse, scroll, time_on_page, etc."
          },
          "timestamp_ms": {
            "type": "integer",
            "title": "Timestamp Ms",
            "description": "Client-side timestamp in milliseconds"
          },
          "payload": {
            "type": "object",
            "title": "Payload",
            "description": "Event-specific data"
          }
        },
        "type": "object",
        "required": [
          "type",
          "timestamp_ms"
        ],
        "title": "TelemetryEvent",
        "description": "Single telemetry event."
      },
...
      "UserAchievementOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "awardedAt": {
            "type": "string",
            "format": "date-time",
            "title": "Awardedat"
          },
          "badge": {
            "$ref": "#/components/schemas/BadgeOut"
          }
        },
        "type": "object",
        "required": [
          "id",
          "awardedAt",
          "badge"
        ],
        "title": "UserAchievementOut"
      },
...
      "UserOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "fullName": {
            "type": "string",
            "title": "Fullname"
          },
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "role": {
            "$ref": "#/components/schemas/Role"
          },
          "nim": {
            "type": "string",
            "title": "Nim"
          },
          "kelas": {
            "type": "string",
            "title": "Kelas"
          },
          "tahunMasuk": {
            "type": "integer",
            "title": "Tahunmasuk"
          },
          "avatarUrl": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avatarurl"
          },
          "zenPoints": {
            "type": "integer",
            "title": "Zenpoints",
            "default": 0
          },
          "currentLvl": {
            "type": "integer",
            "title": "Currentlvl",
            "default": 1
          },
          "lifeMotto": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lifemotto"
          },
          "achievements": {
            "items": {
              "$ref": "#/components/schemas/UserAchievementOut"
            },
            "type": "array",
            "title": "Achievements",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "id",
          "fullName",
          "email",
          "role",
          "nim",
          "kelas",
          "tahunMasuk"
        ],
        "title": "UserOut"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token obtained from /auth/login or /auth/register"
      }
    }
  },
  "security": [
    {
      "BearerAuth": []
    }
  ]
}